requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
coverage==7.13.2
mypy_extensions==1.1.0
pip-check==3.2.1
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
terminaltables==3.1.10
typing_extensions==4.15.0
urllib3==2.6.3
vulture==2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
